Minor BC Break / Bug Fix: finalize() will only be called once, after filters are applied
Filters were already only applied once, but finalizing the query builder would happen every time handleRequest() was called, which was not the original intended behavior. Now finalizing the query builder (including calling the optional finalize() method) will only happen once. As a side effect of this change, finalize() will only ever be called prior to any database queries (never after the count query, as was the case before when not using the Doctrine Paginator). In any case, this is considered minor because, while it's technically possible to call handleRequest() more than once, doing so is not likely to be useful and has never been encouraged.